Implement memory pattern in ScoreDisplay component
- Added state tracking to ScoreDisplay to avoid redundant hardware calls.
- Updated LedControl mock with call counters for verification.
- Added test_ScoreDisplay_Optimization to verify hardware call skipping.
- Documented changes in COMPONENT_LIBRARIES.md.
- Reverted out-of-scope changes and ensured compatibility with existing tests.

+++ src/farkle/test/mocks/src/ScoreDisplay.cpp
@@ -5,16 +5,16 @@ ScoreDisplay::ScoreDisplay(int dataPin, int clkPin, int csPin) {}
 void ScoreDisplay::begin() {}
 
 void ScoreDisplay::addDisplay(DisplayType type, int deviceIndex) {
-    device_map[type] = deviceIndex;
+    device_map[(int)type] = deviceIndex;
 }
 
 void ScoreDisplay::print_number(int number, DisplayType type, bool blink) {
-    captured_numbers[type] = number;
-    captured_blinks[type] = blink;
-    cleared_displays[type] = false;
+    captured_numbers[(int)type] = number;
+    captured_blinks[(int)type] = blink;
+    cleared_displays[(int)type] = false;
 }
 
 void ScoreDisplay::clear(DisplayType type) {
-    cleared_displays[type] = true;
-    captured_numbers[type] = 0; // Or some indicator
+    cleared_displays[(int)type] = true;
+    captured_numbers[(int)type] = 0; // Or some indicator
 }
